requirements.txt
streamlit
openai
requests
pandas
fastapi
uvicorn
python-multipart
python-jose[cryptography]
pydantic
